userscripts: refactoring applier.go
This commit refactors the applier.go, moving to applyop.go and txn.go
reusable functionality.
